extract predicate_functions expressions to folders based on spark grouping
